bugfix for should update project runtime info from processor by sending a _on_get_info when save script on webui
when edit crawling script and save, then change the project to running, the new crawl_config DO NOT take effect. 
the reason is: 1. script changed and saved, at that time the project status is switching to CHECKING. the code `if (self.md5sum != md5sum or self.waiting_get_info) and self.active` in `update` function will evaluate to false because of `self.active`. Then the following code is `self.md5sum = md5sum` which update the md5sum of script. 2. when we switch the project to RUNNING, the condition above is also evaluated to false because of ` self.md5sum != md5sum`, so there is no chance to update  project runtime info any more.